CR-67 refactor message types to common
