+++ pom.xml
@@ -125,7 +125,7 @@
     <dependency>
       <groupId>uk.gov.ons.ctp.integration.common</groupId>
       <artifactId>framework</artifactId>
-      <version>0.0.19</version>
+      <version>0.0.21-SNAPSHOT</version>
     </dependency>
 
     <dependency>
@@ -138,7 +138,7 @@
     <dependency>
       <groupId>uk.gov.ons.ctp.integration</groupId>
       <artifactId>contactcentreserviceapi</artifactId>
-      <version>0.0.33-SNAPSHOT</version>
+      <version>0.0.33</version>
     </dependency>
 
     <dependency>
