+++ src/main/java/uk/gov/ons/ctp/integration/contactcentresvc/event/impl/ContactCentreEventPublisherImpl.java
@@ -1,9 +1,9 @@
-package uk.gov.ons.ctp.integration.contactcentresvc.message.impl;
+package uk.gov.ons.ctp.integration.contactcentresvc.event.impl;
 
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.integration.annotation.MessageEndpoint;
-import uk.gov.ons.ctp.integration.contactcentresvc.message.ContactCentreEventPublisher;
+import uk.gov.ons.ctp.integration.contactcentresvc.event.ContactCentreEventPublisher;
 
 /**
  * Implementation for publication of contact centre asynchronous requests to the Response Management
