--- src/main/java/uk/gov/ons/ctp/integration/contactcentresvc/message/model/Event.java
@@ -1,3 +1,0 @@
-package uk.gov.ons.ctp.integration.contactcentresvc.message.model;
-
-public interface Event {}
