+++ src/main/java/uk/gov/ons/ctp/integration/contactcentresvc/service/EventService.java
@@ -1,11 +1,11 @@
 package uk.gov.ons.ctp.integration.contactcentresvc.service;
 
 import uk.gov.ons.ctp.common.error.CTPException;
-import uk.gov.ons.ctp.integration.contactcentresvc.message.model.Event;
+import uk.gov.ons.ctp.common.event.model.GenericEvent;
 
 /** Service responsible for creation of events */
 public interface EventService {
 
   /** Create and publish event */
-  void createEvent(Event request) throws CTPException;
+  void createEvent(GenericEvent request) throws CTPException;
 }
