+++ src/main/java/uk/gov/ons/ctp/integration/contactcentresvc/service/impl/EventServiceImpl.java
@@ -4,9 +4,9 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uk.gov.ons.ctp.common.error.CTPException;
+import uk.gov.ons.ctp.common.event.model.GenericEvent;
 import uk.gov.ons.ctp.common.jackson.CustomObjectMapper;
-import uk.gov.ons.ctp.integration.contactcentresvc.message.ContactCentreEventPublisher;
-import uk.gov.ons.ctp.integration.contactcentresvc.message.model.Event;
+import uk.gov.ons.ctp.integration.contactcentresvc.event.ContactCentreEventPublisher;
 import uk.gov.ons.ctp.integration.contactcentresvc.service.EventService;
 
 /** Implementation for creation of Event messages */
@@ -19,7 +19,7 @@ public class EventServiceImpl implements EventService {
 
   /** Create and publish event */
   @Override
-  public void createEvent(Event event) throws CTPException {
+  public void createEvent(GenericEvent event) throws CTPException {
     try {
       String message = objectMapper.writeValueAsString(event);
       publisher.sendEvent(message);
